--- apps/meteor/app/message-mark-as-unread/client/index.ts
@@ -1,1 +1,0 @@
-import './useMarkAsUnreadMessageAction';
